refactor(gptme-rag-mcp): rename package from gptme-rag to gptme-rag-mcp
Avoids name collision with the existing gptme/gptme-rag PyPI package
(ChromaDB vector search). The new name clarifies the package's role:
RAG (BM25 book/session search) exposed via MCP protocol.

- packages/gptme-rag/ → packages/gptme-rag-mcp/
- module gptme_rag → gptme_rag_mcp throughout
- All 4 tests pass with renamed imports
