+++ packages/gptme-rag-mcp/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "gptme-rag"
+name = "gptme-rag-mcp"
 version = "0.1.0"
 description = "RAG-as-MCP Knowledge Server for gptme agents — search CS books and session history via MCP"
 authors = [
@@ -40,14 +40,14 @@ test = [
 ]
 
 [project.scripts]
-rag-mcp-server = "gptme_rag.mcp_server:main"
+rag-mcp-server = "gptme_rag_mcp.mcp_server:main"
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/gptme_rag"]
+packages = ["src/gptme_rag_mcp"]
 
 [tool.ruff]
 line-length = 100
