+++ packages/gptme-rag-mcp/src/gptme_rag_mcp/__init__.py
@@ -1,0 +1,3 @@
+"""gptme-rag-mcp: RAG-as-MCP Knowledge Server for gptme agents."""
+
+__version__ = "0.1.0"
