--- packages/gptme-rag/src/gptme_rag/__init__.py
@@ -1,3 +1,0 @@
-"""gptme-rag: RAG-as-MCP Knowledge Server for gptme agents."""
-
-__version__ = "0.1.0"
